+++ lib/StaticNarrative/StaticNarrativeImpl.py
@@ -5,7 +5,11 @@ import os
 from StaticNarrative.exporter.exporter import NarrativeExporter
 from StaticNarrative.uploader.uploader import upload_static_narrative
 from StaticNarrative.narrative_ref import NarrativeRef
-from StaticNarrative.narrative.narrative_util import save_narrative_url
+from StaticNarrative.narrative.narrative_util import (
+    save_narrative_url,
+    get_static_info,
+    verify_admin_privilege
+)
 
 #END_HEADER
 
@@ -27,7 +31,7 @@ class StaticNarrative:
     ######################################### noqa
     VERSION = "0.0.1"
     GIT_URL = "https://github.com/briehl/StaticNarrative"
-    GIT_COMMIT_HASH = "c05ca10c0d11a33157f4b9f028de6c02fae72692"
+    GIT_COMMIT_HASH = "c773a0c8e98a683de627d335abe48d4b6f7d8eed"
 
     #BEGIN_CLASS_HEADER
     #END_CLASS_HEADER
@@ -59,13 +63,22 @@ class StaticNarrative:
         # ctx is the context object
         # return variables are: output
         #BEGIN create_static_narrative
+
+        # init - get NarrativeRef and Exporter set up
         ref = NarrativeRef.parse(params["narrative_ref"])
+        ws_url = self.config["workspace-url"]
+        verify_admin_privilege(ws_url, ctx["user_id"], ctx["token"], ref.wsid)
         exporter = NarrativeExporter(self.config, ctx["user_id"], ctx["token"])
+
+        # set up output directories
         output_dir = os.path.join(self.config["scratch"], str(ref.wsid), str(ref.objid), str(ref.ver))
         os.makedirs(output_dir, exist_ok=True)
+
+        # export the narrative to a file
         output_path = exporter.export_narrative(ref, output_dir)
+        # upload it and save it to the Workspace metadata before returning the url path
         static_url = upload_static_narrative(ref, output_path, self.config["static-file-root"])
-        save_narrative_url(self.config, ctx["token"], ref, static_url)
+        save_narrative_url(self.config["workspace-url"], ctx["token"], ref, static_url)
         output = {
             "static_narrative_url": static_url
         }
@@ -77,6 +90,35 @@ class StaticNarrative:
                              'output is not type dict as required.')
         # return the results
         return [output]
+
+    def get_static_narrative_info(self, ctx, params):
+        """
+        :param params: instance of type "GetStaticNarrativeInfo" ->
+           structure: parameter "ws_id" of type "ws_id" (a workspace id)
+        :returns: instance of type "StaticNarrativeInfo" (ws_id - the
+           workspace id narrative_id - the id of the narrative object made
+           static narrative_version - the version of the narrative object
+           saved url - the url of the static narrative (just the path, the
+           Narrative front end should provide the host) narr_saved - ms since
+           epoch of when the narrative that was made static was saved.
+           static_saved - ms since epoch of when the static narrative was
+           saved) -> structure: parameter "ws_id" of type "ws_id" (a
+           workspace id), parameter "narrative_id" of Long, parameter
+           "narrative_version" of Long, parameter "url" of type "url",
+           parameter "narr_saved" of Long, parameter "static_saved" of Long
+        """
+        # ctx is the context object
+        # return variables are: info
+        #BEGIN get_static_narrative_info
+        info = get_static_info(self.config["workspace-url"], ctx["token"], params.get("ws_id"))
+        #END get_static_narrative_info
+
+        # At some point might do deeper type checking...
+        if not isinstance(info, dict):
+            raise ValueError('Method get_static_narrative_info return value ' +
+                             'info is not type dict as required.')
+        # return the results
+        return [info]
 
     def status(self, ctx):
         """