+++ test/mocks.py
@@ -5,12 +5,11 @@ import json
 from copy import deepcopy
 
 
-def _mock_adapter(ws_id: int = None,
-                  ref_to_file: Dict[str, str] = {},
+def _mock_adapter(ref_to_file: Dict[str, str] = {},
                   ref_to_info: Dict[str, List] = {},
                   ws_info: List = [],
                   user_map: Dict[str, str] = {},
-                  static_nar_ver: int = None):
+                  ws_perms: Dict[int, Dict[str, str]] = {}):
     """
     Sets up mock calls as a requests_mock adapter function.
     Mocks POST calls to:
@@ -19,6 +18,15 @@ def _mock_adapter(ws_id: int = None,
         Workspace.get_object_info3
     Mocks GET calls to:
         Auth (api/V2/users)
+    :param ref_to_file: dict - maps from a workspace ref to the path to a file containing the
+        object JSON. used in Workspace.get_objects2
+    :param ref_to_info: dict - maps from a workspace ref to a object info list for that object.
+        Used in Workspace.get_object_info3
+    :param ws_info: list - the Workspace info that should be returned on a call to
+        Workspace.get_workspace_info
+    :param ws_perms: dict - a mapping from a workspace id to a dict with user permissions on that ws
+    :param user_map: dict - a mapping from user id to full name, used in calls to Auth
+        GET api/V2/users
     """
 
     workspace_meta = dict()
@@ -73,6 +81,9 @@ def _mock_adapter(ws_id: int = None,
                 # updated metadata.
                 ws_id = params[0].get("wsi", {}).get("id")
                 workspace_meta[ws_id] = params[0].get("new")
+            elif method == "Workspace.get_permissions":
+                ws_id = params[0].get("id")
+                result = [ws_perms.get(ws_id, {})]
             response._content = bytes(json.dumps({
                 "result": result,
                 "version": "1.1"
@@ -113,15 +124,15 @@ def _get_object_from_file(filename: str):
 
 
 def set_up_ok_mocks(rqm,
-                    ws_id: int,
-                    ref_to_file: Dict,
-                    ref_to_info: Dict,
-                    ws_info: List,
-                    user_map: Dict):
-    rqm.add_matcher(_mock_adapter(ws_id=ws_id,
-                                  ref_to_file=ref_to_file,
+                    ref_to_file: Dict = {},
+                    ref_to_info: Dict = {},
+                    ws_info: List = [],
+                    ws_perms: Dict = {},
+                    user_map: Dict = {}):
+    rqm.add_matcher(_mock_adapter(ref_to_file=ref_to_file,
                                   ref_to_info=ref_to_info,
                                   ws_info=ws_info,
+                                  ws_perms=ws_perms,
                                   user_map=user_map))
 
 
@@ -148,3 +159,23 @@ def mock_ws_info(ws_id):
 def mock_ws_info_unauth(ws_id):
     pass
 
+
+def mock_ws_bad(rqm, msg):
+    """
+    Always returns a 500 from a workspace call, triggering a ServerError
+    """
+    def mock_adapter_bad_ws(request):
+        response = requests.Response()
+        response.status_code = 500
+        response._content = bytes(json.dumps({
+            "error": {
+                "code": -32500,
+                "message": msg,
+                "error": "long Java vomit stacktrace",
+                "name": "JSONRPCError"
+            },
+            "version": "1.1",
+            "id": request.json().get("id", "12345")
+        }), "UTF-8")
+        return response
+    rqm.add_matcher(mock_adapter_bad_ws)