+++ RELEASE_NOTES.md
@@ -1,6 +1,15 @@
 # StaticNarrative release notes
 =========================================
 
+0.0.2
+-----
+* Added status function. Returns the status of the service, including running version and hash string for the deployed Git commit.
+* Added get_static_narrative_info function. Given a workspace id, return information about the static narrative created from it, if there is any.
+
+0.0.1
+-----
+* Added create_static_narrative function. Takes in a full Narrative reference (including version) and creates a static version of it, hosted on a KBase server.
+
 0.0.0
 -----
 * Module created by kb-sdk init